[InMethod Grid] Post Cherry Pick compilation fixes. Committing before merging latest master from WicketStuff/Core
Cherry Picked into Core 1.5.x Insertable Changes (Mid cherry pick commit)

Added classes:
 + CheckBoxPropertyColumn        - checkbox style column for boolean fields
 + DropDownChoiceColumn          - Drop Down Selector Column
 + DatePropertyColumn            - Editbable property Column for dates
 + AddRecordsToolbar             - button for adding new rows(records) to the datagrid
 + IAppendableDataSource         - datasource interface that supports adding data
 + AppendableDataProviderAdapter - apadter for DataProviders to IAppandableDatasource

Added supporting Methods:
 + DataGrid#insertRow(T rowData)
 + DataGridBody#insertRow(IModel<T> rowModel)
 + DataGridBody#createItem(index, IModel<T>)
 + DataGridBody#clearCache()
 + AbstractPageableview#clearCache

Added Objects
 + trashcan-delete icon
 + NOTICE file with licensing info for trashcan Delete Icon

Added Java doc comments

Fixed EditableColumns, Added DropDownChoiceColumn, Added IAppendableDataSource & AppendableDataProviderAdapter
IAppendableDataSource & AppendableDataProviderAdapter facilitate the back-end of data additions
Converted the Editable columns to extend EditablePropertyColumn
Simplified DatePropertyColumn by having it take DateConverter as an option
Added the ability to clear the cache to AbstractPageableView (necessary for immediate grid update on add)
Simplified some If statements
manually finished the Form Patch
Added Logging to DataGrid
Moved the insert code from DatagridBody to Datagrid
HTML formatting
Misc Code formatting
Corrected JavaDoc

Conflicts:

	jdk-1.5-parent/inmethod-grid-parent/inmethod-grid/src/main/java/com/inmethod/grid/column/CheckBoxColumn.java
	jdk-1.5-parent/inmethod-grid-parent/inmethod-grid/src/main/java/com/inmethod/grid/column/editable/CheckBoxPanel.html
	jdk-1.5-parent/inmethod-grid-parent/inmethod-grid/src/main/java/com/inmethod/grid/column/editable/CheckBoxPanel.java
	jdk-1.5-parent/inmethod-grid-parent/inmethod-grid/src/main/java/com/inmethod/grid/column/editable/CheckBoxPropertyColumn.java
	jdk-1.5-parent/inmethod-grid-parent/inmethod-grid/src/main/java/com/inmethod/grid/column/editable/DatePropertyColumn.java
	jdk-1.5-parent/inmethod-grid-parent/inmethod-grid/src/main/java/com/inmethod/grid/column/editable/DateTextFieldPanel.html
	jdk-1.5-parent/inmethod-grid-parent/inmethod-grid/src/main/java/com/inmethod/grid/column/editable/DateTextFieldPanel.java
	jdk-1.5-parent/inmethod-grid-parent/inmethod-grid/src/main/java/com/inmethod/grid/column/editable/EditableCellPanel.java
	jdk-1.5-parent/inmethod-grid-parent/inmethod-grid/src/main/java/com/inmethod/grid/column/editable/EditablePropertyColumn.java
	jdk-1.5-parent/inmethod-grid-parent/inmethod-grid/src/main/java/com/inmethod/grid/column/editable/TextFieldPanel.java
	jdk-1.5-parent/inmethod-grid-parent/inmethod-grid/src/main/java/com/inmethod/grid/common/AbstractGrid.html
	jdk-1.5-parent/inmethod-grid-parent/inmethod-grid/src/main/java/com/inmethod/grid/common/AbstractGrid.java
	jdk-1.5-parent/inmethod-grid-parent/inmethod-grid/src/main/java/com/inmethod/grid/common/AbstractPageableView.java
	jdk-1.5-parent/inmethod-grid-parent/inmethod-grid/src/main/java/com/inmethod/grid/datagrid/DataGrid.java
	jdk-1.5-parent/inmethod-grid-parent/inmethod-grid/src/main/java/com/inmethod/grid/datagrid/DataGridBody.java
	jdk-1.5-parent/inmethod-grid-parent/inmethod-grid/src/main/java/com/inmethod/grid/toolbar/NoRecordsToolbar.java

Conflicts:

	jdk-1.5-parent/inmethod-grid-parent/inmethod-grid/src/main/java/com/inmethod/grid/column/AbstractColumn.java
	jdk-1.5-parent/inmethod-grid-parent/inmethod-grid/src/main/java/com/inmethod/grid/column/editable/SubmitCancelPanel.java
	jdk-1.5-parent/inmethod-grid-parent/inmethod-grid/src/main/java/com/inmethod/grid/common/ColumnsHeader.java
	jdk-1.5-parent/inmethod-grid-parent/inmethod-grid/src/main/java/com/inmethod/grid/common/ColumnsHeaderRepeater.java
	jdk-1.5-parent/inmethod-grid-parent/inmethod-grid/src/main/java/com/inmethod/grid/common/Icons.java
	jdk-1.5-parent/inmethod-grid-parent/inmethod-grid/src/main/java/com/inmethod/grid/datagrid/DataGridBody.java

+++ jdk-1.6-parent/inmethod-grid-parent/inmethod-grid/src/main/java/com/inmethod/grid/IAppendableDataSource.java
@@ -14,7 +14,7 @@ public interface IAppendableDataSource<T> extends IDataSource<T>
    *  @param index the index value to insert the new Item into
    *  @param item Item to append to the end of the result data
    */
-  public void InsertRow(int index, T item);
+  public void InsertRow(long index, T item);
 
-  public void DeleteRow(int index, T item);
+  public void DeleteRow(long index, T item);
 }
